refactor: remove server module, migrate item-related endpoints to worlds module and events to relay module
