+++ src/routes/relay.py
@@ -1,15 +1,11 @@
-from typing import Annotated
+from fastapi import APIRouter, Response
 
-from fastapi import APIRouter, HTTPException, Response, Query
-
-from src.dependencies.database import db
-from src.models.users import playtime
 from src.models.server import relay
 
-events = APIRouter(prefix='/events', tags=['Events'])
+relay_router = APIRouter(prefix='/relay', tags=['Webhook Relay'])
 
 
-@events.post('/relay', name="Server Relay", status_code=201)
+@relay_router.post('', name="Server Relay", status_code=201)
 async def server_relay_event(body: relay.RelayModel) -> Response:
     """
     Relays a message to the discord server via a webhook.
